third_party/blink/web_tests/virtual/disable-deferred-rendering/fast/canvas/canvas-lost-gpu-context-expected.txt

Test the behavior of canvas recovery after a gpu context loss

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS contextLostTest is false
PASS ctx.isContextLost() is false
PASS ctx.isContextLost() is false
PASS contextLostTest is true
PASS ctx.isContextLost() is true
PASS Graphics context lost event dispatched.
PASS contextLostTest is true
PASS ctx.isContextLost() is true
PASS Context restored event dispatched after context lost.
PASS contextLostTest is false
PASS ctx.isContextLost() is false
